RankingPage: fix crash in XAML designer
When opening the RankingPage in the XAML designer it would often crash
with a null exception. This is because to render the GridView, it would
actually run view model code. In particular it would try to fetch the
grid items via LoadMoreItemsAsync, which would try to call upon
App.Current.Pixiv. This wouldn't be set as it doesn't run the main
App.Xaml.cs code.

If we are running in the designer mode, load a "fake" result instead of
doing the full data/API access. Also put #ifdef DEBUG around this, as we
don't need this check in the production version to avoid slowdown.